src/requirements.txt
asgiref==3.5.2
Django==4.1.4
djangorestframework==3.14.0
mysql-connector==2.2.9
mysqlclient==2.1.1
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
